Rename AboutPageHeader to CustomPageHeader
