--- apps/codeforafrica/src/components/AboutPageHeader/index.js
@@ -1,3 +1,0 @@
-import AboutPageHeader from "./AboutPageHeader";
-
-export default AboutPageHeader;
