+++ apps/codeforafrica/src/components/CustomPageHeader/index.js
@@ -1,0 +1,3 @@
+import CustomPageHeader from "./CustomPageHeader";
+
+export default CustomPageHeader;
